f6/requirements.txt
argparse
biopython>=1.78
icecream>=2.1.0
imbalanced-learn>=0.7.0
joblib>=1.0.0
numpy>=1.19.0
pandas>=1.0.0
pybedtools>=0.8.0
pyBigWig>=0.3.0
pyfaidx>=0.5.0
scikit-learn>=0.24.0
scipy>=1.5.0
statsmodels>=0.11.0
tqdm>=4.50.0
